+++ private-templates-service/client/src/components/Dashboard/RecentlyViewed/OwnerAvatar/index.tsx
@@ -1,0 +1,3 @@
+import OwnerAvatar from './OwnerAvatar';
+
+export default OwnerAvatar;
