+++ private-templates-service/client/src/components/Dashboard/RecentlyViewed/OwnerInfo/styled.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const InfoWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
