+++ private-templates-service/client/src/store/rootReducer.ts
@@ -5,6 +5,7 @@ import { currentTemplateReducer } from "./currentTemplate/reducers";
 import { filterReducer } from "./filter/reducer";
 import { allTemplateReducer } from "./templates/reducer";
 import { recentTemplatesReducer } from "./recentTemplates/reducer";
+import { templateOwnerReducer } from "./templateOwner/reducers";
 
 import { combineReducers } from "redux";
 import { sortReducer } from "./sort/reducer";
@@ -17,7 +18,8 @@ export const rootReducer = combineReducers({
   filter: filterReducer,
   sort: sortReducer,
   allTemplates: allTemplateReducer,
-  recentTemplates: recentTemplatesReducer
+  recentTemplates: recentTemplatesReducer,
+  templateOwner: templateOwnerReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
